# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BurgerPlanet",
        "version": "1.0"
    },
    "paths": {
        "/rest/items": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Obtener todos los items",
                "operationId": "3a43f915d39f44a089f3d3c06613e4f5",
                "responses": {
                    "200": {
                        "description": "Lista de items"
                    }
                }
            }
        },
        "/rest/items/{id}": {
            "get": {
                "tags": [
                    "Items"
                ],
                "summary": "Obtener un item por ID",
                "operationId": "cfe276fce050c74554885f0f95226ac3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del item",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item encontrado"
                    },
                    "404": {
                        "description": "Item no encontrado"
                    }
                }
            }
        },
        "/rest/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Obtener todas las órdenes",
                "operationId": "cd37eabfd98b49bc2a3b194ebc2876dd",
                "responses": {
                    "200": {
                        "description": "Lista de órdenes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Crear una nueva orden",
                "operationId": "b86915443e8e30703ba190da72f99821",
                "requestBody": {
                    "description": "Datos de la orden",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_email",
                                    "total_amount",
                                    "items"
                                ],
                                "properties": {
                                    "customer_email": {
                                        "type": "string",
                                        "example": "juan@gmail.com"
                                    },
                                    "total_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 4567
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "number",
                                                    "example": 2
                                                },
                                                "individual_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 920
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Orden creada exitosamente"
                    }
                }
            }
        },
        "/rest/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Obtener una orden por ID",
                "operationId": "3216e29cff05dbd7b164dac23516ae73",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la orden",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orden encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Orden no encontrada"
                    }
                }
            }
        },
        "/rest/categories": {
            "get": {
                "tags": [
                    "Product Categories"
                ],
                "summary": "Obtener todas las categorías de productos",
                "operationId": "11a4c1f9d8cc195620505d3ded702057",
                "responses": {
                    "200": {
                        "description": "Lista de categorías de productos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductCategory"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rest/categories/{id}": {
            "get": {
                "tags": [
                    "Product Categories"
                ],
                "summary": "Obtener una categoría de producto por ID",
                "operationId": "afe6404e395bf3e2698482afd66e9a39",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría de producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoría encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductCategory"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    }
                }
            }
        },
        "/rest/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Obtener todos los productos",
                "operationId": "4a015e0a6aa705316d79ab26148eaacf",
                "responses": {
                    "200": {
                        "description": "Lista de productos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rest/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Obtener un producto por ID",
                "operationId": "97749e88af1af59bff63bbe996ee61e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Item": {
                "required": [
                    "id",
                    "quantity",
                    "individual_price",
                    "order_id",
                    "product_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "individual_price": {
                        "type": "number",
                        "format": "float",
                        "example": 12.34
                    },
                    "order_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "product_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "Order": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "customer_email": {
                        "type": "string",
                        "example": "carlos@gmail.com"
                    },
                    "total_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 123.45
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Item"
                        }
                    }
                },
                "type": "object"
            },
            "Product": {
                "required": [
                    "id",
                    "name",
                    "price",
                    "stock",
                    "description",
                    "product_category_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Fanta"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 999.99
                    },
                    "stock": {
                        "type": "integer",
                        "example": 10
                    },
                    "description": {
                        "type": "string",
                        "example": "Vaso de 300ml de Fanta."
                    },
                    "product_category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category": {
                        "$ref": "#/components/schemas/ProductCategory"
                    }
                },
                "type": "object"
            },
            "ProductCategory": {
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Bebida"
                    }
                },
                "type": "object"
            }
        }
    }
}